// src/users/dto/postinga.dto.ts
export class PostinganDto {
  nama_instansi: string;
  nama_pekerjaan: string;
  bidang_pekerjaan: string;
  penuh_waktu: string;
  paruh_waktu: string;
  magang: string;
  kontrak: string;
  harian: string;
  provinsi: string;
  kabupaten: string;
  kecamatan: string;
  deskripsi: string;
  gaji: string;
  skils: string;
  minimal_pendidikan: string;
  pengalaman_kerja: string;
}